+++ src/App.js
@@ -1,43 +1,23 @@
-import { useState } from 'react';
+
 import './App.css';
-import { Link } from "react-router-dom";
+import { Link } from 'react-router-dom';
+import Chatbox from './Chatbox';
+import FriendList from './FriendList';
 
 function App() {
 
-const [text,setText] = useState("")
-
-const [lines,setLines] = useState(["msgl"]);
-
-const onTextChange = (event)=>{
-  setText(event.target.value);
-};
-
-const onSend = () => {
-  setLines(lines => [...lines,text]);
-  setText("");
-};
 return (
   
     <div className="App">
       <div className="App-header">
-      <Link to="/friendList" className="Button">
+      <Link to="/friendList">
                 <button>
                     Go FriendList
                 </button>  
             </Link>
       </div>
-      <div className="App-chatroom">
-        {
-          lines.map(x =>{
-          return <div className="App-chatroom-text">
-                 {x}
-               </div>
-          })
-        }
-      </div>
-      <div className="App-textbox">
-        <input type="text" className="App-textbox-input" value={text} onChange={onTextChange}/>
-        <div className="App-textbox-send"onClick={onSend}>Send!</div>
+      <div className="App-content">
+      <Chatbox/>
       </div>
     </div>
   );
