+++ src/project/ProChatroom.js
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import './ProChatroom.css';
-import { Link } from "react-router-dom";
+import { Link } from 'react-router-dom';
 
 
 function ProChatroom() {
@@ -19,37 +19,33 @@ const onSend = () => {
 };
 return (
   
-    <div className="App">
-      <p className = "noelly">noel</p>
-      <div className="App-header">
-      <div className="circlesbox1">
+  <div className="App">
+    <div className="header">
+       <div className="noelly">noel</div> 
+      <div className="noelphoto">
         <Link to='/ProProfile'>
-          <button>
-            <img src="5.jpg" width="50px" height="50px"/>
-          </button>
-        </Link> 
-       </div>
-       <div className="imp">
-        <Link to='/ProProfile'>
-          <button>
-            <img src="333.jpg" width="50px" height="50px"/>
-          </button>
+            <img src="Logo noel.jpg" width="50px" height="50px"/>
         </Link>
        </div>
-       <div className="imp2"><img src= "5555.jpg"width="60px" height="60px" /></div>
+       <div className="imp">
+        <Link to=''>
+            <img src="333.jpg" width="50px" height="50px"/>
+        </Link>
+       </div>
+       <div className="imp2"><img src= "5555.jpg"width="50px" height="50px" /></div>
       </div>
-      <div className="App-chatroom">
+      <div className="chatroom">
         {
           lines.map(x =>{
-          return <div className="App-chatroom-text">
+          return <div className="chatroom-text">
                  {x}
                </div>
           })
         }
       </div>
-      <div className="App-textbox">
-        <input type="text" className="App-textbox-input" value={text} onChange={onTextChange}/>
-        <div className="App-textbox-send"onClick={onSend}><img src= "send.png" alt = "send.png" width="70px" height="70px"/></div>
+      <div className="textbox">
+        <input type="text" className="textbox-input" value={text} onChange={onTextChange}/>
+        <div className="textbox-send"onClick={onSend}><img src= "send.png" alt = "send.png" width="70px" height="70px"/></div>
       </div>
     </div>
   );
